add compression to static js and css files

+++ prototypes/sitebuilder/management/commands/build.py
@@ -3,7 +3,7 @@ import shutil
 
 from django.conf import settings
 from django.core.management import call_command
-from django.core.management.base import BaseCommand
+from django.core.management.base import BaseCommand, CommandError
 from django.core.urlresolvers import reverse
 from django.test.client import Client
 
@@ -19,12 +19,27 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         # request pages and build output
-        if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
-            shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
-        os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
-        os.makedirs(settings.STATIC_ROOT)
+        settings.DEBUG = False
+        settings.COMPRESS_ENABLED = True #enable compression with django-compressor
+        if args:
+            pages = args
+            available = list(get_pages())
+            invalid = []
+            for page in pages:
+                if page not in available:
+                    invalid.append(page)
+            if invalid:
+                msg = 'Invalid pages: {}'.format(', '.join(invalid))
+                raise CommandError(msg)
+        else:
+            pages = get_pages()
+            if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
+                shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
+            os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
+            os.makedirs(settings.STATIC_ROOT)
         call_command('collectstatic', interactive=False, clear=True, verbosity=0)
-        client=Client()
+        call_command('compress', interactive=False, force=True) #add compression during build
+        client = Client()
         for page in get_pages():
             url = reverse('page', kwargs={'slug': page})
             response = client.get(url)
@@ -33,5 +48,7 @@ class Command(BaseCommand):
             else:
                 output_dir = os.path.join(settings.SITE_OUTPUT_DIRECTORY, page)
                 os.makedirs(output_dir)
+                if not os.path.exists(output_dir):
+                    os.makedirs(output_dir)
             with open(os.path.join(output_dir, 'index.html'), 'wb') as f:
                 f.write(response.content)
